backend/requirements.txt
# Data Handling & Scientific Computing
pandas==2.2.2
numpy==1.26.4

# Machine Learning & AI
scikit-learn==1.5.0
xgboost==2.0.3
torch==2.3.1
transformers==4.41.2
shap==0.45.1

# Database Connection
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9  # For PostgreSQL

# API Interaction
requests==2.32.3
fredapi==0.5.2
newsapi-python==0.2.7

# Configuration
python-dotenv==1.0.1

pandas-ta==0.3.14b
setuptools
